# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects/{projectId}/devices/register": {
      "post": {
        "description": "\n      Registers a new device for push notifications with enhanced validation.\n      Automatically detects platform and validates device tokens.\n      Rate limited to 100 requests per minute.\n    ",
        "operationId": "register",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "example": "64f1a2b3c4d5e6f7a8b9c0d1",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "64f1a2b3c4d5e6f7a8b9c0d3",
                    "projectId": "64f1a2b3c4d5e6f7a8b9c0d1",
                    "token": "fcm_token_here_1234567890",
                    "platform": "android",
                    "userId": "user123",
                    "isActive": true,
                    "tags": [
                      "premium",
                      "beta-tester"
                    ],
                    "properties": {
                      "appVersion": "1.2.3",
                      "deviceModel": "Pixel 7",
                      "osVersion": "13"
                    },
                    "lastSeen": "2023-12-07T16:45:00.000Z",
                    "createdAt": "2023-12-07T16:45:00.000Z",
                    "updatedAt": "2023-12-07T16:45:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid device data or token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Register device",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices": {
      "get": {
        "operationId": "findAll",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "platform",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topics",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/register/basic": {
      "post": {
        "operationId": "registerBasic",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/validate-token": {
      "post": {
        "description": "\n      Validates a device token for a specific platform.\n      Checks token format and verifies with the respective push service.\n      Rate limited to 10 requests per minute due to external API calls.\n    ",
        "operationId": "validateToken",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Token validation result",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "valid": true,
                    "platform": "android",
                    "tokenType": "fcm",
                    "lastValidated": "2023-12-07T16:50:00.000Z",
                    "metadata": {
                      "appId": "com.example.app",
                      "projectId": "firebase-project-123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid token format"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Validate device token",
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/validate-tokens-batch": {
      "post": {
        "operationId": "validateTokensBatch",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/cleanup-tokens": {
      "post": {
        "operationId": "cleanupTokens",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dryRun",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/refresh-token": {
      "post": {
        "operationId": "refreshToken",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/{id}/token": {
      "patch": {
        "operationId": "updateToken",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/stats": {
      "get": {
        "operationId": "getStats",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/tags": {
      "get": {
        "operationId": "getProjectTags",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/tags/stats": {
      "get": {
        "operationId": "getTagStats",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/properties/stats": {
      "get": {
        "operationId": "getPropertyStats",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/segment/query": {
      "post": {
        "operationId": "queryDevicesBySegment",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/segment/count": {
      "post": {
        "operationId": "countDevicesBySegment",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/segment/tokens": {
      "post": {
        "operationId": "getTokensBySegment",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/v1/projects/{projectId}/devices/segment/test": {
      "post": {
        "operationId": "testSegmentQuery",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SegmentQuery"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
...
    "/api/v1/projects/{projectId}/devices/{id}/tags": {
      "post": {
        "operationId": "addTags",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      },
      "delete": {
        "operationId": "removeTags",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
...
  },
  "info": {
    "title": "Push Notification Service API",
    "description": "\n      Enterprise-grade Push Notification Service supporting iOS, Android, and Web platforms.\n      \n      ## Features\n      - 🚀 Multi-platform push notifications (FCM, APNs, Web Push)\n      - 📱 Advanced device management and segmentation\n      - 📊 Real-time analytics and delivery tracking\n      - 🏷️ Template-based notifications with variable substitution\n      - ⚡ High-performance queue system with BullMQ\n      - 🔒 Enterprise security with API key authentication\n      - 🛡️ Rate limiting for API protection\n      - 📈 99.9% uptime reliability target\n      \n      ## Authentication\n      All endpoints require an API key in the `X-API-Key` header.\n      \n      ## Rate Limiting\n      API endpoints are rate-limited to ensure fair usage:\n      - **High Frequency**: 100 requests/minute (notifications, events)\n      - **Medium Frequency**: 300 requests/minute (device management)\n      - **Low Frequency**: 1000 requests/hour (analytics, admin)\n      \n      Rate limit headers are included in all responses:\n      - `X-RateLimit-Limit`: Maximum requests allowed\n      - `X-RateLimit-Remaining`: Requests remaining in current window\n      - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Push Notification Service Team",
      "url": "https://github.com/your-repo/push-notification-service",
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for project authentication. Get your API key from the project dashboard."
      }
    },
    "schemas": {
...
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {}
      },
...
      "SegmentQuery": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}